pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
version = "0.3.1.0000"  # NOTE: This is a placeholder; CI overwrites it based on git tags + commit count. It's set only to the last manual push.
name = "braingeneers"
authors = [
  { name = "UCSC Braingeneers", email = "braingeneers-admins-group@ucsc.edu" },
]
maintainers = [
  { name = "David Parks", email = "dfparks@ucsc.edu" },
  { name = "Kate Voitiuk", email = "kvoitiuk@ucsc.edu" },
]
description = "Braingeneers Python utilities"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering",
  "Typing :: Typed",
]
dependencies = [
  "anyio>=4,<5",
  "awswrangler==3.*",
  "boto3==1.35.95",  # until https://github.com/boto/boto3/issues/4398 is fixed
  "braingeneers-smart-open==2023.10.6",
  "deprecated",
  "h5py",
  "httpx>=0.28,<1",
  "mcp>=1.26,<2",
  "matplotlib",
  "nptyping",
  "numpy",
  "paho-mqtt>=2",
  "pandas",
  "PyJWT[crypto]>=2.10,<3",
  "PyYAML>=6,<7",
  "redis",
  "requests",
  "schedule",
  "scipy",
  "spikedata",
  "tenacity",
  "typing_extensions>=4.6; python_version<'3.11'",
  "diskcache",
  "pytz",
  "tzlocal",
]

[project.optional-dependencies]
all = [
  "braingeneers[ml]",
  "braingeneers[dev]",
]
ml = [
  "torch",
  "scikit-learn",
]
mcp = []
dev = [
  "pytest >=6",
  "pytest-cov >=3",
  "sphinx>=4.0",
  "myst_parser>=0.13",
  "sphinx_book_theme>=0.1.0",
  "sphinx_copybutton",
  "sphinx_autodoc_typehints",
  "furo",
  "joblib",
]

[dependency-groups]
dev = [
    "braingeneers[all]",
]

[project.urls]
Homepage = "https://github.com/braingeneers/braingeneerspy"
"Bug Tracker" = "https://github.com/braingeneers/braingeneerspy/issues"
Discussions = "https://github.com/braingeneers/braingeneerspy/discussions"
Changelog = "https://github.com/braingeneers/braingeneerspy/releases"

[tool.hatch]
metadata.allow-direct-references = true
envs.default.dependencies = [
  "pytest",
  "pytest-cov",
]

[tool.hatch.build.targets.wheel]
packages = ["src/braingeneers"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = [
  "error",
  "ignore::DeprecationWarning"
]
log_cli_level = "INFO"
testpaths = [
  "tests",
]


[tool.coverage]
run.source = ["braingeneers"]
port.exclude_lines = [
  "pragma: no cover",
  "\\.\\.\\.",
  "if typing.TYPE_CHECKING:",
]

[tool.mypy]
files = ["src", "tests"]
python_version = "3.10"
warn_unused_configs = true
strict = true
show_error_codes = true
enable_error_code = ["ignore-without-code", "redundant-expr", "truthy-bool"]
warn_unreachable = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = "braingeneers.*"
disallow_untyped_defs = true
disallow_incomplete_defs = true


[tool.ruff]
select = [
  "E", "F", "W", # flake8
  "B",           # flake8-bugbear
  "I",           # isort
  "ARG",         # flake8-unused-arguments
  "C4",          # flake8-comprehensions
  "EM",          # flake8-errmsg
  "ICN",         # flake8-import-conventions
  "ISC",         # flake8-implicit-str-concat
  "G",           # flake8-logging-format
  "PGH",         # pygrep-hooks
  "PIE",         # flake8-pie
  "PL",          # pylint
  "PT",          # flake8-pytest-style
  "PTH",         # flake8-use-pathlib
  "RET",         # flake8-return
  "RUF",         # Ruff-specific
  "SIM",         # flake8-simplify
  "T20",         # flake8-print
  "UP",          # pyupgrade
  "YTT",         # flake8-2020
  "EXE",         # flake8-executable
  "NPY",         # NumPy specific rules
  "PD",          # pandas-vet
]
extend-ignore = [
  "PLR",    # Design related pylint codes
  "E501",   # Line too long
]
typing-modules = ["braingeneers._compat.typing"]
src = ["src"]
unfixable = [
  "T20",  # Removes print statements
  "F841", # Removes unused variables
]
exclude = []
flake8-unused-arguments.ignore-variadic-names = true
isort.required-imports = ["from __future__ import annotations"]

[tool.ruff.per-file-ignores]
"tests/**" = ["T20"]
"noxfile.py" = ["T20"]


[tool.pylint]
py-version = "3.10"
ignore-paths = []
reports.output-format = "colorized"
similarities.ignore-imports = "yes"
messages_control.disable = [
  "design",
  "fixme",
  "line-too-long",
  "missing-module-docstring",
  "wrong-import-position",
]

[tool.uv.sources]
braingeneers = { workspace = true }
